Compute the constellation a body is in (#199)
This implements all 88 [IAU designated constellations](https://en.wikipedia.org/wiki/IAU_designated_constellations) and the algorithm to find which one holds a set of equatorial coordinates.

The boundaries have been defined based on the B1875 epoch, which means we need to use astrometric (mean J2000) equatorial coordinates and convert them to the B1875 epoch to match how the constellations have been designed.

For now, a constellation is only made of an abbreviation and a name. In the future, I would like to add information for each of them, but also support other cultures.

## How it works

The data comes from the [VizieR archive](https://cdsarc.u-strasbg.fr/viz-bin/Cat?VI/42) and have been pre-processed to allow fast an easy computation by Astronoby.

The different values of right ascension, declinations and abbreviations have been sorted in `sorted_right_ascensions.dat`, `sorted_declinations.dat` and `indexed_abbreviations.dat`. The combinations are pre-processed in `radec_to_index.dat`, where each row represents a specific right ascension, each column represents a specific declination, and each cell contains an index that points to a constellation abbreviation.

This allows for applying a lookup process with binary search for both right ascension and declination in order to compute the index of the cell in `radec_to_index.dat`, which itself will provide the constellation abbreviation.

From this abbreviation, the last step is to get the associated constellation object from the constellations repository.

The files are not large, it is acceptable to load them only once and keep them in memory, so that only the first calculation may take a little bit more time and the next ones are fast.

This implementation is heavily inspired by [Skyfield](https://github.com/skyfielders/python-skyfield).