Add `#phase_angle` and `#illuminated_fraction` to planets
While generally mostly interesting for the Moon, the phase angle and
illuminated fraction are two figures that can be useful for anticipating
how a planet looks like.

This is particularly the case for Mercury and Venus. As inferior planets
(with orbits closer to the Sun than the Earth), they are never observed
fully illuminated and always show a phase.

For superior planets, the information is less interesting as we
constantly see them almost completely illuminated. But still, the
information is there.

Because the phase angle depends on the vector between the sun and the
observer, the sun needs to be computed for each planet instance. This
isn't great, however the new feature of caching should help.