Apparent magnitude (#201)
Magnitude is a measure of the luminosity. The more luminous an object, the lower its magnitude number.

The apparent magnitude represents the brightness of the object seen from Earth. Note that is doesn't account the atmospheric effect.

Most planets have many different coefficients. It is because there is no single formula for describing the brightness of all the variety of the planets.
Note that some cases are ignored and fall back to a default (less accurate) formula:
- Saturn's rings add significant brightness but taking them into account means calculating their inclination, which is a complex one I don't want to deal with right now
- Uranus uses the default formula as the accurate one depends on a complex correction term called "Uranus' sub-Earth and sub-solar latitudes", which again is just too much burden for now

Values have been tested against the IMCCE, JPL Horizons, Stellarium and Skyfield library. They match these sources most of the time, otherwise the difference is too small to be noticed in any amateur purpose.

Since `phase_angle` has been added to planets, the position of the Sun in different reference frames is necessary for multiple planetary calculations. Therefore, the whole sun object is now saved on the planet for computing different data.