Add `#events_between` to `TwilightCalculator` with better accuracy (#204)
To mimic `RiseTransitSetCalculator`, this adds `#events_between` to `TwilightCalculator`, allowing to compute twilight times for a date range instead of having to loop over each date. This drastically improves the performance over long date ranges.

This was also the opportunity to tune the algorithm in order to get closer to the source of truth for this feature: IMCCE. This change made the difference with IMCCE data for test scenarios to e few seconds instead of between one and two minutes previously.

Finally, `utc_offset` optional parameter was added to `#event_on` to ensure the times are properly calculated for a date of local time. Without this parameter, the times are calculated for happening between midinight and midnight UTC (00:00+00:00), which is for example not the definition of a civil day in Singapore, where a civil day is between midnight and midnight GMT+8 (00:00+08:00). The more far from the Greenwich meridian, the more likely the algorithm was to compute morning or evening twilight times for the wrong date.