`#angular_diameter` on Solar System bodies (#207)
Moves `#angular_diameter` from the `topocentric` position to the Solar System body.

This is meant to harmonise with other physical methods like `#apparent_magnitude` or `#phase_angle` which are also defined on the Solar System body and not on a particular reference frame.

This slightly changes the value we previously got for the Moon, which proximity makes the angular diameter influenced by the parallax angle. But I believe this is fine, to the benefit of a more readable API. If someone wants the exact topocentric apparent angular diameter, the calculation by hand is really accessible.