Rename "epoch" (#208)
The ambiguous "epoch" term has been partially removed in ab37cdde1209621f010b4b755fd91cf24e11edd5 in favour of "Julian Date".

This intends to finish this work. The term "epoch" remains in the codebase for equatorial coordinates because it seems it makes more sense in the literature than "instant" or "time", equatorial coordinates are always defined for an epoch.

Multiple methods have been changed:

* `MeanObliquity`: `::for_epoch` to `::at` and now takes an `instant` as parameter
* `MeanObliquity`: `::obliquity_of_reference_in_milliarcseconds` to `::obliquity_of_reference_in_arcseconds`
* `TrueObliquity`: `::for_epoch` to `::at` and now takes an `instant` as parameter
* `Coordinates::Equatorial`: `#to_ecliptic` method now takes an `instant` as parameter