Calculate apoapsis and periapsis events (#213)
As Solar System bodies have elliptical orbits (and not circular), there are moments when they are closer or further from their primary body (Sun for planets, Earth for the Moon).

The further point on the orbits is called the apoapsis, while the closer is the periapsis. Each body reach them once during each period. Technically, when the Sun is the primary body, they are called aphelion and perihelion, while when the primary is the Earth they are called apogee and perigee. Astronoby uses the generic term for all bodies.

This change introduces the ability to compute when these points are reached and how far from the primary body the body is. It searches between two provided dates.
Note that is is possible that no event happen during the provided time range. For example, the Moon's orbital period is ~27 days, which means looking for an event within an hour is very unlikely to result in finding one.

The result is an array of `MaximumEvent` which exposes `#instant` (`Astronoby::Instant`) and `#value` (`Astronoby::Distance`).

```rb
# All aphelions happening for Saturn between 2015 and 2050

ephem = Astronoby::Ephem.load("inpop19a.bsp")

calculator = Astronoby::ExtremumCalculator.new(
  ephem: ephem,
  body: Astronoby::Saturn,
  primary_body: Astronoby::Sun
)

events = calculator.apoapsis_events_between(Time.utc(2015, 1, 1), Time.utc(2051, 1, 1))

events.each do |event|
  puts "#{event.instant.to_time.round} (#{event.value.km.round} km)"
end
# 2018-04-17 11:27:15 UTC (1505798987 km)
# 2047-07-15 06:57:26 UTC (1502883334 km)
```