Deep Sky Object: Handle velocities properly
In the initial implementation, the astrometric and apparent velocities
were off as they were not geocentric but represented the object's
intrinsic velocity.

This is fixed now to match the fact that in Astronoby, astrometric and
apparent reference frames are geocentric.

Note: apparent velocity is slightly off compared to Skyfield and
Astropy. This is because Astronoby applies precession and nutation
corrections to the velocity, which apparently Skyfield and Astropy do
not. This was kept this way as this makes sense for planets.