Deep Sky Object: difference between the body and the position
In the same way using `Astronoby::Jupiter.new` enables to get the
position of the planet in multiple reference frames for a given instant, it
is necessary to provide an API as convenient for deep-sky objects.

Previously, it was only possible to instantiate the body and its
position at the same time, having to provide all the information at
once. Not only this makes things less convenient to get the position at
different instants, it also requires the code to know about all the
attributes constantly. In a next PR we will handle using deep-sky
objects in services like the `RiseTransitSetCalculator` and the current
implementation of `DeepSkyObject` makes things hard.

Now, with a difference between the body (`DeepSkyObject`) and the
position (`DeepSkyObjectPosition`), it is possible to manipulate the
body accross the code and instantiate it (get a position) whenever
necessary.